Refactor testing utilities
