+++ src/test-helpers/registries/index.ts
@@ -1,0 +1,2 @@
+export * from './kusamaRegistry';
+export * from './polkadotRegistry';
