+++ src/test-helpers/registries/polkadotRegistry.ts
@@ -1,0 +1,30 @@
+import { Metadata, TypeRegistry } from '@polkadot/types';
+import { getSpecTypes } from '@polkadot/types-known';
+
+import { polkadotV16MetadataRpc } from '../metadata/polkadotV16Metadata';
+
+/**
+ * Create a type registry for Polkadot.
+ * Useful for creating types in order to facilitate testing.
+ */
+function createPolkadotRegistry(): TypeRegistry {
+	const registry = new TypeRegistry();
+
+	registry.register(getSpecTypes(registry, 'Polkadot', 'polkadot', 16));
+	registry.setChainProperties(
+		registry.createType('ChainProperties', {
+			ss58Format: 0,
+			tokenDecimals: 12,
+			tokenSymbol: 'DOT',
+		})
+	);
+
+	registry.setMetadata(new Metadata(registry, polkadotV16MetadataRpc));
+
+	return registry;
+}
+
+/**
+ * Polkadot v16 TypeRegistry.
+ */
+export const polkadotRegistry = createPolkadotRegistry();
